+++ src/lib/CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Hidden webparts can now be skipped during transformation #194 [jansenbe - Bert Jansen]
 - Improved Tenant Id fetch method #198 [gautamdsheth - Gautam Sheth]
 - Fix if file is template in /libname/Forms folder and therefore has no ListItemId #206 [czullu - Christian Zuellig]
+- Allow for setting the http client timeout via SharePointPnPHttpTimeout as environment variable or app setting [jansenbe - Bert Jansen]
 
 ## [1.2.0]